@@ -3,143 +3,110 @@
   "private": true,
   "license": "MIT",
   "repository": "github:apollographql/apollo-server",
+  "type": "module",
   "scripts": {
     "clean": "git clean -dfqX -- ./node_modules **/{dist,node_modules}/ **/tsconfig*tsbuildinfo",
+    "precompile": "node scripts/precompile.mjs",
     "compile": "tsc --build tsconfig.build.json",
-    "compile:clean": "tsc --build tsconfig.build.json --clean",
-    "watch": "tsc --build tsconfig.build.json --watch",
-    "release:version-bump": "lerna version",
-    "release:version-bump:server": "npm run release:version-bump -- --force-publish=apollo-server,apollo-server-core,apollo-server-azure-functions,apollo-server-cloud-functions,apollo-server-cloudflare,apollo-server-express,apollo-server-fastify,apollo-server-hapi,apollo-server-koa,apollo-server-lambda,apollo-server-micro,apollo-server-integration-testsuite",
-    "release:start-ci-publish": "node -p '`Publish (dist-tag:${process.env.APOLLO_DIST_TAG || \"latest\"})`' | git tag -F - \"publish/$(date -u '+%Y%m%d%H%M%S')\" && git push origin \"$(git describe --match='publish/*' --tags --exact-match HEAD)\"",
-    "postinstall": "lerna run prepare && npm run compile",
-    "pretest": "tsc --build tsconfig.json",
+    "postcompile": "node scripts/postcompile.mjs",
+    "watch": "tsc --build tsconfig.esm.json --watch",
+    "pretest": "npm run compile",
     "test": "jest --verbose",
-    "test:clean": "jest --clearCache",
     "test:watch": "jest --verbose --watchAll",
-    "testonly": "npm test",
-    "test:ci": "npm run coverage -- --ci --maxWorkers=2  --reporters=default --reporters=jest-junit",
+    "test:smoke": "npm run test:smoke:prepare && npm run test:smoke:run",
+    "test:smoke:prepare": "smoke-test/prepare.sh",
+    "test:smoke:run": "smoke-test/smoke-test.sh",
+    "test:ci": "npm run coverage -- --ci --maxWorkers=2 --reporters=default --reporters=jest-junit",
+    "test:one-integration": "npm run test -- packages/server/src/__tests__/express4/integration.test.ts",
     "coverage": "npm test -- --coverage",
     "coverage:upload": "codecov",
     "codegen": "graphql-codegen",
-    "prettier-check": "prettier --check .",
-    "spell-check": "cspell lint '**' --no-progress || (echo 'Add any real words to cspell-dict.txt.'; exit 1)"
+    "codegen-check": "(npm run codegen && npm run compile) || (echo 'Current schema types are incompatible with existing usage. Update generated types with `npm run codegen` and resolve any TypeScript errors.'; exit 1)",
+    "lint": "eslint .",
+    "prettier-check": "prettier -v && prettier --check .",
+    "prettier-fix": "prettier -v && prettier --write .",
+    "spell-check": "cspell lint '**' '.changeset/**' --no-progress || (echo 'Add any real words to cspell-dict.txt.'; exit 1)",
+    "changeset-publish": "npm run compile && changeset publish",
+    "changeset-check": "changeset status --verbose --since=origin/main",
+    "changeset-version": "changeset version && npm i"
   },
+  "workspaces": [
+    "packages/*"
+  ],
   "engines": {
-    "node": ">=12",
-    "npm": "8.x"
-  },
-  "dependencies": {
-    "@apollographql/apollo-tools": "0.5.2",
-    "apollo-datasource": "file:packages/apollo-datasource",
-    "apollo-datasource-rest": "file:packages/apollo-datasource-rest",
-    "apollo-reporting-protobuf": "file:packages/apollo-reporting-protobuf",
-    "apollo-server": "file:packages/apollo-server",
-    "apollo-server-azure-functions": "file:packages/apollo-server-azure-functions",
-    "apollo-server-cache-memcached": "file:packages/apollo-server-cache-memcached",
-    "apollo-server-cache-redis": "file:packages/apollo-server-cache-redis",
-    "apollo-server-caching": "file:packages/apollo-server-caching",
-    "apollo-server-cloud-functions": "file:packages/apollo-server-cloud-functions",
-    "apollo-server-cloudflare": "file:packages/apollo-server-cloudflare",
-    "apollo-server-core": "file:packages/apollo-server-core",
-    "apollo-server-env": "file:packages/apollo-server-env",
-    "apollo-server-errors": "file:packages/apollo-server-errors",
-    "apollo-server-express": "file:packages/apollo-server-express",
-    "apollo-server-fastify": "file:packages/apollo-server-fastify",
-    "apollo-server-hapi": "file:packages/apollo-server-hapi",
-    "apollo-server-integration-testsuite": "file:packages/apollo-server-integration-testsuite",
-    "apollo-server-koa": "file:packages/apollo-server-koa",
-    "apollo-server-lambda": "file:packages/apollo-server-lambda",
-    "apollo-server-micro": "file:packages/apollo-server-micro",
-    "apollo-server-plugin-base": "file:packages/apollo-server-plugin-base",
-    "apollo-server-plugin-response-cache": "file:packages/apollo-server-plugin-response-cache",
-    "apollo-server-types": "file:packages/apollo-server-types"
+    "node": ">=14.16.0",
+    "npm": ">=8.5.0"
   },
   "devDependencies": {
-    "@apollo/client": "3.5.10",
-    "@graphql-codegen/cli": "2.6.2",
-    "@graphql-codegen/typescript-operations": "2.3.5",
-    "@graphql-tools/mock": "8.6.1",
-    "@graphql-tools/schema": "8.3.3",
-    "@graphql-tools/utils": "8.6.3",
-    "@hapi/hapi": "20.2.1",
-    "@josephg/resolvable": "1.0.1",
-    "@rollup/plugin-commonjs": "21.0.2",
-    "@rollup/plugin-json": "4.1.0",
-    "@sinonjs/fake-timers": "9.1.1",
-    "@types/async-retry": "1.4.3",
-    "@types/aws-lambda": "8.10.93",
-    "@types/body-parser": "1.19.2",
-    "@types/bunyan": "1.8.8",
-    "@types/connect": "3.4.35",
-    "@types/cors": "2.8.12",
-    "@types/express": "4.17.13",
-    "@types/express-serve-static-core": "4.17.28",
-    "@types/fast-json-stable-stringify": "2.0.0",
-    "@types/hapi__hapi": "20.0.10",
-    "@types/ioredis": "4.28.10",
-    "@types/jest": "27.4.1",
-    "@types/koa-router": "7.4.4",
-    "@types/lodash": "4.14.180",
-    "@types/lodash.sortby": "4.7.6",
-    "@types/lodash.sumby": "4.6.6",
-    "@types/lodash.xorby": "4.7.6",
-    "@types/lru-cache": "5.1.1",
-    "@types/memcached": "2.2.7",
-    "@types/micro": "7.3.7",
-    "@types/node": "12.20.47",
-    "@types/node-fetch": "2.6.1",
-    "@types/qs-middleware": "1.0.1",
-    "@types/request": "2.48.8",
-    "@types/request-promise": "4.1.48",
-    "@types/sinonjs__fake-timers": "8.1.1",
-    "@types/supertest": "2.0.12",
-    "@types/test-listen": "1.1.0",
-    "@types/type-is": "1.6.3",
-    "@types/uuid": "8.3.4",
-    "@vendia/serverless-express": "4.5.4",
+    "@apollo/client": "3.12.9",
+    "@apollo/gateway": "2.9.3",
+    "@apollo/subgraph": "2.9.3",
+    "@apollo/utils.createhash": "2.0.2",
+    "@changesets/changelog-github": "0.5.0",
+    "@changesets/cli": "2.27.12",
+    "@graphql-codegen/cli": "3.3.1",
+    "@graphql-codegen/typescript-operations": "3.0.4",
+    "@graphql-tools/mock": "8.7.20",
+    "@graphql-tools/schema": "9.0.19",
+    "@rollup/plugin-commonjs": "28.0.2",
+    "@types/async-retry": "1.4.9",
+    "@types/compression": "1.7.5",
+    "@types/cors": "2.8.17",
+    "@types/express": "4.17.21",
+    "@types/express-serve-static-core": "4.19.6",
+    "@types/lodash": "4.17.15",
+    "@types/lodash.sumby": "4.6.9",
+    "@types/lodash.xorby": "4.7.9",
+    "@types/negotiator": "0.6.3",
+    "@types/node": "14.18.63",
+    "@types/node-fetch": "2.6.12",
+    "@types/qs-middleware": "1.0.4",
+    "@types/retry": "0.12.5",
+    "@types/superagent": "8.1.9",
+    "@types/supertest": "6.0.2",
+    "@types/test-listen": "1.1.2",
+    "@types/type-is": "1.6.7",
+    "@types/uuid": "9.0.8",
+    "@types/whatwg-mimetype": "3.0.2",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
+    "@vendia/serverless-express": "4.12.6",
     "awaiting": "3.0.0",
-    "azure-functions-ts-essentials": "1.3.2",
-    "body-parser": "1.19.2",
-    "bunyan": "1.8.15",
     "codecov": "3.8.3",
-    "connect": "3.7.0",
-    "cspell": "5.19.2",
-    "deep-freeze": "0.0.1",
-    "express": "4.17.3",
-    "fastify": "3.27.4",
-    "form-data": "4.0.0",
-    "graphql": "15.8.0",
+    "compression": "1.7.5",
+    "cors": "2.8.5",
+    "cspell": "6.31.3",
+    "eslint": "8.57.1",
+    "eslint-plugin-import": "2.31.0",
+    "express": "4.21.2",
+    "graphql": "16.10.0",
+    "graphql-subscriptions": "3.0.0",
     "graphql-tag": "2.12.6",
-    "ioredis": "4.28.5",
-    "jest": "27.5.1",
-    "jest-config": "27.5.1",
-    "jest-junit": "13.0.0",
+    "jest": "29.7.0",
+    "jest-config": "29.7.0",
+    "jest-junit": "16.0.0",
+    "jest-mock": "29.7.0",
     "jest-mock-random": "1.1.1",
-    "js-sha256": "0.9.0",
-    "koa": "2.13.4",
-    "koa-router": "10.1.1",
-    "lerna": "4.0.0",
+    "jest-serializer-html": "7.1.0",
     "lodash.sumby": "4.6.0",
-    "log4js": "6.4.4",
-    "memcached-mock": "0.1.0",
-    "mock-req": "0.2.0",
-    "nock": "13.2.4",
-    "node-fetch": "2.6.7",
-    "prettier": "2.6.0",
+    "nock": "13.5.6",
+    "node-fetch": "2.7.0",
+    "prettier": "3.4.2",
     "qs-middleware": "1.0.3",
-    "request-promise": "4.2.6",
     "requisition": "1.7.0",
-    "rollup": "2.70.1",
-    "supertest": "6.2.2",
+    "rollup": "3.29.5",
+    "supertest": "7.0.0",
     "test-listen": "1.1.0",
-    "ts-jest": "27.1.3",
-    "typescript": "4.6.2",
-    "winston": "3.6.0",
-    "winston-transport": "4.5.0"
+    "ts-jest": "29.2.5",
+    "typescript": "5.4.5"
   },
   "jest": {
     "projects": [
       "<rootDir>/packages/*/jest.config.js"
     ]
+  },
+  "volta": {
+    "node": "22.13.1",
+    "npm": "9.9.4"
   }
 }
