@@ -1,6 +1,7 @@
-const { defaults } = require('jest-config');
+import { defaults } from 'jest-config';
+import { createRequire } from 'node:module';
 
-module.exports = {
+export default {
   testEnvironment: 'node',
   setupFilesAfterEnv: ['<rootDir>/../../jest.setup.js'],
   preset: 'ts-jest',
@@ -9,10 +10,17 @@ module.exports = {
   testPathIgnorePatterns: ['/node_modules/', '/dist/'],
   moduleFileExtensions: [...defaults.moduleFileExtensions, 'ts', 'tsx'],
   clearMocks: true,
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/src/__tests__/tsconfig.json',
-      diagnostics: false,
-    },
+  transform: {
+    '^.+\\.test.ts$': [
+      'ts-jest',
+      {
+        tsconfig: '<rootDir>/src/__tests__/tsconfig.json',
+        diagnostics: true,
+      },
+    ],
+  },
+  moduleNameMapper: {
+    // Ignore '.js' at the end of imports; part of ESM support.
+    '^(\\.{1,2}/.*)\\.js$': '$1',
   },
 };
